refactor(layout): extract shared Header and Footer using Route Group (landing)
