+++ src/app/(landing)/pricing/page.tsx
@@ -10,8 +10,6 @@ import { request } from "@/lib/utils/request";
 import { MODULES } from "@/constants/modules";
 import { CREDIT_PLANS } from "@/config/credit_plans";
 import PricingCard from "@/components/pricing/pricing_card";
-import Header from "@/components/landing_page/header";
-import Footer from "@/components/landing_page/footer";
 import { Check, Minus, Plus, Lock, Loader2 } from "lucide-react";
 import ConfirmModal from "@/components/common/confirm_modal";
 import AuthModal from "@/components/auth/auth_modal";
@@ -205,7 +203,6 @@ export default function PricingPage() {
 
   return (
     <div className="bg-white">
-      <Header />
 
       <main className="isolate">
         <div className="relative pt-14 text-center sm:pt-20 lg:pt-32">
@@ -696,7 +693,6 @@ export default function PricingPage() {
         />
       </main>
 
-      <Footer />
     </div>
   );
 }
